feat(storage): 完善示例元数据和数据浏览页
变更内容：
- 新增并调整交易空间示例元数据，补充本地平台默认配置
- 统一分页 total 类型处理，并补充生成代码测试
- 优化数据浏览/概览页面，支持中文数据集名、数据ID列表、分页表格和固定列展示

自测：
- cd modules/storage && go test ./...
- cd modules/control && go test ./...
- cd web && pnpm test:unit
- cd web && pnpm build:prod

+++ modules/control/internal/service/control/service.go
@@ -345,5 +345,5 @@ func pageSlice[T any](items []T, page *pb.Page) ([]T, *pb.PageResult) {
 	if end > len(items) {
 		end = len(items)
 	}
-	return items[start:end], &pb.PageResult{Page: pageNo, Size: size, Total: uint64(len(items)), HasMore: end < len(items)}
+	return items[start:end], &pb.PageResult{Page: pageNo, Size: size, Total: uint32(len(items)), HasMore: end < len(items)}
 }
