+++ modules/storage/internal/services/access/service.go
@@ -941,7 +941,7 @@ func pageSlice[T any](items []T, page *pb.Page) ([]T, *pb.PageResult) {
 	if end > len(items) {
 		end = len(items)
 	}
-	return items[start:end], &pb.PageResult{Page: pageNo, Size: size, Total: uint64(len(items)), HasMore: end < len(items)}
+	return items[start:end], &pb.PageResult{Page: pageNo, Size: size, Total: uint32(len(items)), HasMore: end < len(items)}
 }
 
 func archiveFileOverlaps(item *pb.ArchiveFile, timeRange *pb.TimeRange) bool {
